Render updates and refactoring
- Refactored render, moved functions to a separate file.
- Changed clouds generation.
- Changed chunk generation, added grass top to dirt blocks.
- Updated lighting and SAO settings.
